Updates tests to work with phpunit-mock-objects > 4
Prior to this commit, when running `composer install` the locked version of
phpunit would install phpunit@6.4.3 and phpunit-mock-objects@^4.0.3

After running `composer update` in order to get the changes for `illuminate\support` to take effect
the installed version of phpunit is 6.5.14 which brings in phpunit-mock-objects@^5.0.9

The `PHPUnit_Framework_MockObject_Invocation_Object` was changed to `PHPUnit\Framework\MockObject\Invocation\ObjectInvocation`
between versions 4 and 5 of phpunit-mock-object. The `ObjectInvocation` class does not have a public property of `$methodName`,
but instead has a private property of `$methodName` with a get method. The same goes for parameters.